requirements.txt
flask
flask-cors
pandas
requests
beautifulsoup4
duckduckgo-search
openpyxl
gunicorn
redis
google-generativeai
python-dotenv